// src/app/shared/app/models/MasterTables/production/i-libraries-filter.ts
export interface ILibrariesFilter {
	sNo: number | null;
	type: string | null;
	defaultTick: number | null;
	class: string | null;
	lineOfBusiness: string | null;
	insuranceCopmany: null;
	item: string | null;
	itemArabic: string | null;
	description: string | null;
	descriptionArabic: string | null;
	identity: string | null;
}